[rdct] IJPL-197945: Move registering disposing logic of the composite
Since listening for disposing of the composite logic has been registered in `bindComposite` function, handling of the composite's closing in the cases when no editors have been found, is badly managed.

That means that if the composite fails to open, it could be lost in limbo — it will stay opened on the backend, but on the frontend it will be closed and that could not be changed.

This commit moves logic to the earlier stage to handle all possible shenanigans with fails

GitOrigin-RevId: e2017c39e5df73ad7b341718edea688f77a70b7d