[JEWEL-948] Add Scrollable containers for staggered grids
This adds a new overload to ScrollableContainer APIs that takes in a
`ScrollableState`. This new API allows it to be used with any
scroll-based container, both lazy and non-lazy.

This then deprecates the older lazy-based overloads as they can be
expressed trivially in terms of the new ScrollableState-based ones.

Lastly, this adds clarification to the various KDocs.