[JEWEL-948] Add Scrollable containers for staggered grids
This adds a new overload to ScrollableContainer APIs that takes in a
`ScrollableState`. This new API allows it to be used with any
scroll-based container, both lazy and non-lazy.

This then deprecates the older lazy-based overloads as they can be
expressed trivially in terms of the new ScrollableState-based ones.

Lastly, this adds clarification to the various KDocs.

closes https://github.com/JetBrains/intellij-community/pull/3166

GitOrigin-RevId: 72032b66f20eb30c2f12f25a5c9bf41c6a1d0b92