[JEWEL-944] Annotate breaking changes in API dumps
This adds a script that's run by the GitHub CI to validate that the
PR is not introducing binary changes, and if it finds any, it adds
annotations that show on the PR code.

This script should be able to be used locally as well, as when there
is no PR_NUMBER env set, it just checks against HEAD~1 (assumes our
changes are always at most 1 commit).