Rename Mac distro dir to 'android-studio'
We don't want to rebuild the platform when switching from Canary to
Stable, so we cannot rely on 'isEAP' (from platform build scripts)
to determine the Mac bundle name. Instead we will adjust the Mac
bundle name later in our Bazel build.

For consistency with Linux/Windows, we'll use 'android-studio' as the
placeholder Mac bundle directory.

Bug: 338090219
Test: existing, e.g. //tools/adt/idea/studio:test_studio
Change-Id: If69a35939c6d5c0c44f79089aad0f014f5e6735d