IJPL-188102 Eel+WSL: register handlers for WSL files as early as possible
This functionality had worked a long time ago but was broken after refactorings. We expected that reverting the commit dae3a216 should resolve all problems, but it turned out that it's not enough.

In the master branch this code is completely refactored. In general, it works in master the same way as with this commit, but with a much cleaner architecture.

GitOrigin-RevId: 1c1e3a940f2e420e9f277abfcdd9b62ce3bc20d1