IJPL-196857 Introduce initOnShow
The problem with launchOnShow was its contract.
At first, it wouldn't even start the job if the component
is hidden immediately and then shown again.
Then it was fixed, but it still wouldn't restart the job
if it's cancelled before completing.
But the most common use case for this function
was to initialize UI components using a suspending API.
Which means that if the component is hidden before
it's initialized and then shown again, it would remain
uninitialized forever.

The replacement function targets these use cases
specifically, hence the name.
It's intended to be used with coroutines that complete
normally within sensible time and must be restarted
if they were cancelled before that.

GitOrigin-RevId: 6b46f7d19e0c104ec585b30dcf15b89efffbbeff