[terminal] IJPL-200524 Key event handling refactoring
Rename register/unregister because for the new
Override IDE Shortcuts implementation they may not always
actually register the dispatcher.

Make the dispatcher private for the same reason:
it may not even be needed when the option is off.
Rename setupKeyEventDispatcher to not mention
the dispatcher.

Rename "skipped" actions to "allowed."
The wording was very confusing:
a "skipped" action meant that we need to skip its
key event and not process it in the terminal
because it belongs to an action allowed to execute.

GitOrigin-RevId: 257b30e8113a322ce57b2d06601ecf7ce8b0b23b