[JEWEL-943] Create InfoText API
This introduces a simple InfoText API that allows users to easily show
info-styled text. This text uses the medium typography, and the info
text colour.

This also adds KDoc to the `Text` composables, mostly taken from the
Material `Text` composables, which have the same signature as ours,
plus a bunch of tweaks.

closes https://github.com/JetBrains/intellij-community/pull/3172

GitOrigin-RevId: ee9b0d265ec4dc8711e10e11b6b25ee2828e316d