[JEWEL-943] Create InfoText API
This introduces a simple InfoText API that allows users to easily show
info-styled text. This text uses the medium typography, and the info
text colour.

This also adds KDoc to the `Text` composables, mostly taken from the
Material `Text` composables, which have the same signature as ours,
plus a bunch of tweaks.