[terminal] IJPL-200549 Fix absence of close confirmation dialog on tab closing
Since we now create `TerminalTabCloseListener` before adding the content to the ContentManager, the listener was not added.
Change it in the platform to always add the listener for content manager property change. So, the required `ContentManagerListener` will be added once ContentManager is set.
Now `BaseContentCloseListener` can be installed before adding the content to the manager.

GitOrigin-RevId: b2113c06f9b4853e0917f3ab102bf17c8a0e513b