IJPL-49424 Fix the DragListener.SHARED leak
The problem: the listener is retrieved using reflection
through TransferHandler.getDropTargetListener.
The returned object has a "support" field
which holds a strong reference to the component inside.
There's no way to clear it without resorting to reflection,
because it only can be modified using the setDNDVariables
function, which is private and is never called with a null value.

The solution: let's just create a new listener instead.
It's cached in the data context, but thankfully it's easy
to clear because the key is it's class, which is available to us.

GitOrigin-RevId: 9e909b749eb9a1055466135b5f3a4ef272a536ea