IJPL-49424 Run the Swing leak fixes when reusing the project frame
Previously they were run only when closing the frame.
That was mostly OK because those leaks would quickly
fix themselves if the frame was reused for a new project.
But now there's this smooth auto scroll listener,
which isn't triggered that often, so it's best to clear
it up anyway when the frame is reused.
Doing the other leak fixes there won't harm either.

GitOrigin-RevId: 0fb3dc43773c876892bab8d9d50a9a7a7eb3a077