[plugin model tests] report an error if an embedded or required content module depends on an optional module from the same plugin
Such configurations are confusing because the optional module actually becomes required. Also, if an embedded module depends on an optional module, this will cause a circular dependency between classloaders.
Dependencies on optional 'intellij.platform.backend' are temporarily not reported until IJPL-201428 is fixed (they don't cause problems for now because there are no classes in the module).

GitOrigin-RevId: dfbbb69102046f30db8c32e0e0de531bd8d1c5a0