[vcs] IJPL-187179 Make the Close Commit Dialog action work only for the Commit tool window
We could've just checked the tool window ID
in getCommitDialog(), but because of a BGT action update,
we shouldn't really use the tool window API at all,
including ToolWindow.type, as it's not thread safe.

So instead we just provide the result as a part of the data context.
But we still have to check the window ID when we do so,
or else the action would affect other VCS tool windows,
for example, the Git tool window.

GitOrigin-RevId: 030218a40d93420d6f5135d43120e5b45c13a85d