PY-83287 Fix an infinite recursion during name resolution of property accessor decorators
During resolving decorators like `@x.setter`, `x` was checked for overloads.
For that, every function, including the one originally decorated by `@x.setter` was checked for having an `@overload` decorator.
For that, `@x.setter` was resolved again, leading to an infinite recursion.

To fix that, a direct check is added to return early if `setter`, `getter` or `deleter` are converted to known decorators.

GitOrigin-RevId: 2e9eaccd3146b17734902b6be69ebe7d9e9a8b95