[air] lsp: group lsp markup into a single entity
LSP markup, in particular, produced by the call to obtain semantic
tokens, is rather massive in quantity. This lead to a bunch of errors
related to "very long change" and general degradation of performance.

Now all the markup is a set of items in a single entity per LSPDocument
(documentMarkup).

In addition, exceeding the retries due to ContentModified error will no
longer throw a hard exception.

Instead, we will throw a cancellation exception, with the expectation
that the caller will retry his request (or give up, if he's not
interestend anymore) at a more appropriate time, which the caller should
know better, e.g. by using a debouncer or some sort of synchronisation
to verify that the document changes have settled and it's sensisble to
try again.

GitOrigin-RevId: 19ba398a7b4a39aab441d3fadc66ccfa917c4ef3