IJPL-201073 Synchronize pane loading in the Project View
One problem was that the code was only partially synchronized.
This could've led to a variety of issues.
For example, when setupImpl() calls ensurePanesLoaded(),
it might return immediately because isExtensionsLoaded
is true, but because ensurePanesLoaded() itself
was not synchronized, it could've been possible
that some other thread is executing it right now.
So it would be possible that isExtensionsLoaded is true,
but the panes were not actually loaded yet.

Another problem was that getExtensions()
is in theory cancellable, especially on the first invocation.
If the first invocation happens in a cancellable action
update (through getSelectInTargets), then it could throw
an exception, leaving isExtensionsLoaded in the true state,
thus making sure that the panes will never be loaded at all.

Fix both issues by synchronizing the entire pane loading process
and by introducing isExtensionsLoading, so now we can
both prevent recursion by setting it immediately and allow
for a second loading attempt if the previous one fails
(so isExtensionsLoaded remains false).

Take care to make sure that the code following
isExtensionsLoaded = true is exception-safe.

GitOrigin-RevId: ef8a4760a03180777f8ae535d534421948af6fb8