jpsbootstrap: skip noisy messages on teamcity
during a single build this verbose logging can
produce 100mb of logs purely for initiated and
processed messages. this was partially handled
in VMT-MR-3078 for vm-templates, but we also
have these data being logged in jpsbootstrap
steps of regular builds. the information is
not that valuable compared to the amount of
logs it produces and started/completed should
be enough to investigate problematic
artifacts.

GitOrigin-RevId: b97b277fff4a9b17cc74e81769a5c4814ec3f8a9