IJPL-195829 [recent files] Refactor the factory API
Introduce separate overloads for the switcher and recent files,
to hide the implementation detail parameters,
and to make the expectation on other parameters clearer
(for example, that event is null for the recent files,
and onlyEditedFiles is null for the switcher).

This refactoring will help adding a key released handler
needed to fix the issue with events being missed
while suspending code is running.
Because this is only an issue for the switcher,
not for the recent files popup.

GitOrigin-RevId: 3a9b40a2191c563b01721c65767384ab0906a92b