IJPL-195829 [recent files] Fix the issue with missing key release events
Because of the suspending code, some EDT events may
happen between actionPerformed() and the actual showing
of the switcher. If the "modifier released" event happens
to be one of those, the switcher will never receive it
and will never close.

To fix, capture all "modifier released" events between
the initial action event and the actual showing.
Then forward all those to the key release listener.
Do it in a separate event (using launch) just in case,
to make sure everything else is processed,
so the switcher is in the "up and running" state.

GitOrigin-RevId: 92cee293fb4794fcc7c3c34531e613c5c6da5463