[build scripts] API: deprecate ProductModulesLayout::excludedModuleNames
If modules which are included in the core classloader are excluded via this property, it may lead to errors at runtime if other modules depend on them. Also, it may cause compatibility issues with plugin which use classes from these modules. If a module isn't present in all products based on the IntelliJ platform, it should be registered as an optional content module, and all other plugins and modules which depend on it should declare explicit dependency in their XML descriptors. It's possible to remove some of such modules from the distribution via ContentModuleFilter. Currently, it's supported for the frontend distribution (JetBrains Client).

GitOrigin-RevId: f2ee380a6bf713695fe46c3435e54bbb1f378ed1