IJPL-184662 Moved contributors to the backend for now
For some reason they are conflicting in JavaRefactoringTest RD test somehow.

If I try to reproduce the scenario in split dev - works just fine.

But in test environment:
- injection present
- navigation declaration works
- declaration/usages are not highlighted and on attempt to refactor - getting missing declaration error

(cherry picked from commit 324f65768497336fc9fa6a926da5f756d6c7ce15)

This is the second attempt. And the symptom is now different - `This Action is not implemented in JetBrains Client`

GitOrigin-RevId: 6a76237a3e6eae7cd08623b19c25c9b1e2e24792