[terminal] IJPL-202203 Fix the IOOBE in type ahead
When inserting a string, the document might get trimmed.
When it happens, the cursor may end up at the end
of the document already, so attempting to move it might fail.

To make this part reliable and not dependent on the details
of the model implementation, let's use absolute offsets
to compute the new cursor offset. And to be absolutely safe,
let's also coerceIn it to the maximum valid value.

GitOrigin-RevId: 6582b0467cd5419bf66abd07b5406a7a1011ace2