RUBY-34687 include element in reference invalidation for completion
When invoking completion in the editor, a fake file and element is created with the text "IntelliJIDEARulezzzz". These fake file/elements do not bump the PSI modification counter, and mean that any attempt to query their reference will retrieve the cached reference, even if the caret has moved inside the element to a different offset.

Using the element as well as the standard PSI modification counter ensures that the cache is cleared in completion when the caret moves offset.

GitOrigin-RevId: 2273fa7fcf8bb38e181fc98b3a41fd8e40bbc7a0