IJPL-161517 Fix editor splitters resize
This code was originally intended to handle only loading decorators,
so it breaks on other layered panes. In particular, there's a layered
pane in the editor since the floating toolbar was added.
Because of this, this code would never return the vertical scrollbar,
as it's a direct child of that layered pane.

Now that it's fixed, JBScrollPane.canBePreprocessed
correctly handles this scrollbar and doesn't let the splitter's
divider to take over as long as the mouse is over the scrollbar's
thumb.

GitOrigin-RevId: 50c0d640f561db87c3879d8fd3e8694693f80675