[gradle][ExternalSystem] IDEA-377941 add ExternalProjectEntity - Workspace entity
This Workspace entity is added to represent a project and store its `externalProjectPath`: e.g., a path of the root project or of a linked project. It is supposed to be helpful in further implementing Workspace entities for the Gradle project model: they will be connected with this entity instead of keeping the `externalProjectPath` in themselves. However, this entity also could be useful for other build tools.

For Gradle, ExternalProjectEntity is created in `GradleContentRootSyncContributorTest` - in the same place where GradleBuildEntity and GradleProjectEntity will be created. This contributor is chosen for locating these entities because it's planned to use GradleProjectEntity as a content root in the future, instead of JPS.

Tests:
* org.jetbrains.plugins.gradle.importing.syncAction.GradleContentRootSyncContributorTest

GitOrigin-RevId: c1937d9d563db23cb82d220c193427b745d28886