[build scripts] load system properties from *.vmoptions and product-info.json file when starting IDE in dev build mode
Before, system properties for the IDE process were manually specified in multiple places. To make IDE running in 'dev build' mode more similar to the production one, they are now loaded from the same files which are used in the production launcher. This way we also don't need to manually specify additional properties for running the frontend process in 'dev build' mode (IJPL-185957).

GitOrigin-RevId: 19dd59d4a30db82a652aa17197c849be1140964f