IJPL-188785 Fix tool window header double click
It was not processed for tool windows with tab DnD
because the tabs would occupy all available width,
leaving no space for the parent component
that actually processed double clicks.

This was introduced when migrating from the MiG layout,
as it also specified grow() for such tool windows,
while for others it was just growY(). However, it turns
out that grow() actually didn't work for some reason,
and the size was effectively the same as usual.

Therefore, fix this by dropping that growFirst
property completely.

A side effect of this change is that now it's a bit harder
to drag the tabs horizontally, as going outside the tab
area will trigger splitting instead, but since the tabs
themselves are usually not very narrow, it doesn't
seem to be a noticeable UX issue. Anyway, it was
like that until 2025.1 and nobody complained about it.

GitOrigin-RevId: 495d1bf13596971e903ff79766b9ac0d957005e3