[JEWEL-985] Fix API issues found with Metalava
This fixes several issues that cropped up thanks to Metalava:
* Correctly annotate `LocalCodeHighlighter` and `NoOpCodeHighlighter`
  as experimental (they expose the experimental highlighter interface)
* Avoid breaking change in `DefaultComponentStyling`
* Avoid breaking changes in `ComponentStyling.dark()` and `.light()`
  in standalone's `IntUiTheme.kt`
* Made `JewelTheme.instanceUuid` and `LocalThemeInstanceUuid` stable
* Properly annotate `LocalPopupRenderer` as experimental
* Deprecate `LocalMenuManager` that exposes deprecated `MenuManager`
* Fixed `JewelApiVersion` folder sourceset type in Gradle build
* Cleaned up a bunch of deprecated API usages

Also, there is a leakage of experimental APIs in stable APIs:
 * `ComboBox` exposes `PopupManager`
 * `EditableComboBox` exposes `PopupManager`
 * `MarkdownText` exposes various Markdown APIs

We'll have to fix in 0.31. It is mostly an annoyance for users, so
it's not urgent enough to warrant a patch in 0.30.0 in my opinion.
The fix is likely going to have to be that we'll make those exposed
APIs stable.

The Metalava setup will be added separately in JEWEL-981, as it is a
very noisy change, and we'll need a custom Metalava Gradle setup — the
Gradle Metalava plugin that's out there does not work well. These API
issues have been spotted, and these changes have been done, thanks to
said Metalava setup, anyway. It's just going to be committed in a
separate PR/commit.

closes https://github.com/JetBrains/intellij-community/pull/3193

(cherry picked from commit dca226dac9dd1dbe748a009fbab66c6c26ce011b)


(cherry picked from commit 5869af3895bf6208035bbb6bf5e0bb610e4bf6be)

IJ-MR-173046

GitOrigin-RevId: 892c13b8789d3595d61381346af88a8e68774ee6