[show usages] IJPL-201422 UsageNavigation for multiple UsageInfos
NavigationService.navigate is using a semaphore with DROP_OLDEST,
which essentially means "the last request wins." But here, it's not
even clear which request is the last one, as they're launched
concurrently using CoroutineScope.launch.

Thankfully, this thing is local and has only one invocation,
so we can easily change it to accept a list of UsageInfos.
Then the implementation is trivial.

One side effect is that now all non-UsageInfo usages are processed
first (in a modal blocking manner), but that shouldn't be much
of an issue, as the order is generally not important as long
as all the usages are opened.


(cherry picked from commit 9706ff825a0ee34c9820192782deaa83915cbd89)

IJ-CR-173097

GitOrigin-RevId: 3c66ba4515d6ac2403d663b407ee4197f326d9f9