[gradle] IDEA-375500 remove the crutch that tries to fix identityPath calculation for buildSrc and Gradle < 8.0
This commit breaks only one thing: identity path calculation for buildSrc belonging to the ROOT build. On the other hand, now the calculation result is the same as Gradle provides in GradleProjectUtil.getProjectIdentityPath.

For Gradle lower than 8.2, identityPath for a project is calculated relying on the known hierarchy of builds and projects. There is a problem with buildSrc hierarchy with Gradle < 8.0: buildSrc is synced separately, so its parent is unknown. Without that, identity path calculation is incorrect.

Previously, I added a crutch in `DefaultGradleLightProject.getBuildIdentityPath` that fixed identityPath calculation if buildSrc belongs to the root build. However, the calculation for buildSrc in included builds was still incorrect. We decided that if the crutch is not strong enough to fix this behavior for buildSrc of both root build and included build - it's better to remove it.

Tests verifying correct behavior:
* org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test identity path for buildSrc of the root build AFTER 8,0
* org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test identity path for buildSrc of an included build AFTER 8,0

Tests related to the broken behavior:
* org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test identity path for buildSrc of the root build BEFORE 8,0
* org.jetbrains.plugins.gradle.model.DefaultGradleLightProjectTest.test identity path for buildSrc of an included build BEFORE 8,0
* org.jetbrains.plugins.gradle.model.DefaultGradleLightBuildTest.test converted buildSrc does not have a parent before Gradle 8,0

GitOrigin-RevId: 81afa1c55efbb76ef733a1412edd831bdf9e0b02