[JEWEL-981, JEWEL-979] Add Metalava signature generator
This adds the Metalava Gradle plugin to the Gradle build. I have also
generated the 0.29.0 and 0.30.0 dumps.

I had originally used an existing Gradle Metalava plugin, but that
didn't seem to work well for us. It especially failed at passing custom
args to Metalava, which we need to do.

Instead, I forked the setup from https://github.com/autonomousapps/dependency-analysis-gradle-plugin
and customised it to better fit our needs:
* Use the latest Metalava
* Fix all the issues I found in it (no idea how it even compiles for
  them, even though we're on the same Gradle version)
* Add a distinction between stable and experimental dumps

This also:
* Updates the IJ icon generator to emit a properly annotated AllIcons
  keys file
* Updates JUnit 5 to the latest version, 5.13.4 [Gradle]
* Fixes the JUnit setup for the detekt-plugin module [Gradle]
* Bumps the IJP target to the latest stable (252.23892.409) [Gradle]
* Removes the old Gradle-based icon generator (JEWEL-979) [Gradle]
* Runs Metalava on the GitHub CI
* Update GitHub CI commit validator to support multiple issues in one

closes https://github.com/JetBrains/intellij-community/pull/3199

GitOrigin-RevId: 8255383d77be24919f91e45ccbe1340c25696429