[jps build devkit] refactoring: extract code which generates the runtime module repository to a separate module (IJPL-201730)
Currently, the runtime module repository is generated by a plugin inside the JPS build during compilation. Doing the same in Bazel doesn't look like a good option because generating something for all modules in the project contradicts the incremental nature of Bazel. So it's better to invoke generation of the runtime module repository by JPS project configuration inside build scripts. To do that and avoid having dependencies on the JPS build classes inside the build scripts, the code which generates the runtime module repository is moved to a separate module 'intellij.devkit.runtimeModuleRepository.generator'.

GitOrigin-RevId: f2596fd51c1c019d304de36ed93fe163cf519ae6