[JEWEL-987] Fix Detekt issues in ide-plugin module
The checks were not being run so a few slipped through. This fixes
them and makes sure some basic Detekt form is run on the GitHub CI
as well, to avoid repeats.

This happened because we can't run detektMain/detektTest as they
require the module to compile, but it can't. IDE plugins need to
reference released builds of the platform, which are bound not to have
the stuff we haven't released yet.

closes https://github.com/JetBrains/intellij-community/pull/3196

GitOrigin-RevId: 90ba2de863423669470f714159e4ccd45b86f196