Build IntelliJ Platform with EAP=false always
We will now build IntelliJ Platform with EAP=false for all Studio
releases (including Canary). This has a few benefits:

* It removes the need to rebuild IntelliJ prebuilts when switching
  from Canary to Stable (facilitating faster, more frequent releases).

* It facilitates building multiple Studio configurations (Canary,
  Stable, etc.) in the same repo branch using the same set of IntelliJ
  prebuilts.

* Conceptually, we prefer to build atop stable IntelliJ releases, and
  any EAP-specific behavior inside the platform tends to be less
  useful to us. For example, extra debugging assertions are useful to
  JetBrains engineers during upstream stabilization, but tend to be
  less useful to us during the Studio Canary cycle.

Any Canary-specific behavior we want to retain will be handled on the
Android plugin side using alternative mechanisms.

Side note: the platform build scripts complain if 'majorReleaseDate'
is not set by AndroidStudioApplicationInfo.xml in a non-EAP build. As
a workaround we'll override it with a nil value using the appropriate
system property.

Bug: 338090219
Test: //tools/adt/idea/studio:test_studio (for example)
Change-Id: Idd52feb7c4a6887bb322c77cabf280dab8695abe