[jps build devkit] refactoring: extract interface ResourcePathsSchema interface (IJPL-201730)
If the JPS build doesn't generate the runtime module repository, we need to produce it in build scripts from scratch. It will be more efficient and convenient to generate it with proper resource paths for modules and libraries directly instead of first generating it with JPS output paths and then replacing the paths.

GitOrigin-RevId: 3c57f41ba719857be1a15032747dc5ae5bddb629