[kotlin] Add bootstrap repo to the intellij build
This way, it is possible to test the build of the intellij repository
against the bootstrap build. Bootstrap builds are released from
master ~every two weeks.

In compiler plugin gradle build files, the `if "dev" in KOTLIN_VERSION`
was used, because this is mostly for testing, and bootstrap repo contains
only dev versions

GitOrigin-RevId: ada056fd627312b69ad7dca79a9488073204081a