[remote dev] never show the dialog about importing config for the IDE backend (IJPL-187164)
If the IDE is running as a remove dev backend, UI cannot be shown before `Application` is initialized because it replaces the standard `awt.toolkit` with com.intellij.platform.impl.toolkit.IdeToolkit which depends on `Application`. Currently, 'idea.initially.ask.config' option is explicitly set to 'never' in launchers for the remote dev backend, but it wasn't set in run configurations for the split mode. So it seems that it's better to have this handled in code.

(cherry picked from commit 3167f76e18e2f675ffd1fd73ce6b992e6b09b497)

GitOrigin-RevId: 2c632fd6b55d427bf51270ffb3ae9bd826934f4b