[JEWEL-994] Add branch comparison script
Add script to compare branches for differences in Jewel commits. The
script can be used to get a list of commits to cherry-pick for a new
release, or to sanity check before publishing that all the commits
that need to be in a branch are there (master vs release branch, or
release branch vs release branch).

Usage: compare-branches [<options>] <branch1> <branch2>

closes https://github.com/JetBrains/intellij-community/pull/3203

GitOrigin-RevId: be49c27da6cd01683adba462f256af26d7e839e8