[remdev] IJPL-166441 Fix restoring the Project View state
1. Because the pane is added externally using addProjectPane,
it would not get its state applied, as that normally happens
only with panes managed by ProjectViewImpl directly.
So add an addProjectPane overload that allows restoring
the state nevertheless.

2. Because remdev nodes don't provide a presentable text
into PresentationData, let's have another piece of heuristics
that tries to guess it by text attributes. Alternatively,
we could add it to the model directly, but it doesn't seem
like a good reason to alter RD models.

GitOrigin-RevId: 4cc660771e7cc8d8370e0f33a0a96f96b53d9773