[lsp] fix `FirDeclaration was not found for class org.jetbrains.kotlin.psi.KtParameter` exception when `kotlin.analysis.partialBodyAnalysis` is enabled
The issue was in the way user data is handled in the Analyzer. The user data handling is delegated to theExternalUserDataStorage which may or may not be shared between the requests. Stubs heavily rely on the DIRECT_PARENT_IS_STUBBED user data key to understand the stub parent. Because the ExternalUserDataStorage was not shared between the indexing request and the highlighting request, this flag was lost. Thus, partial body analysis was not able to find a proper parent of some local declaration, so instead of KtBlockExpression being a parent of some local function, it was a KtNamedFunction that broke the partial body resolve logic of finding the nearest resolvable declaration.

^LSP-253 fixed

GitOrigin-RevId: 6e2bb21800e6cf3df624bc65a094e16eca33fea7