IJPL-200317 Fix lateinit property mainSplitters has not been initialized
We're trying to open editors before ide frame finished its initialization and set `mainSplitters` to a proper value.
We should wait until manager is ready to open files.

The best place to do that seems to be even before we subscribe to composite events as we do it only once per project session. And it's guaranteed that after the initialization files can safely open

GitOrigin-RevId: aee950cbebeac16f4f70fd1fd29e36fde430a027