IJPL-203550 platform: fix thread leak from "Find in Path" popup
The issue is in 'runOnAllThreads': 'runnable.run()' task inherits Job from the context, while
the 'ourExecutor.submit' do not.
As a result, one thread might get cancelled, while leaving the rest spinning.

The fix is to make sure the cancelled threads properly update counters and do not cause the
'runnersCount - 23', 'idleCount - 22', 'deque.size() - 0' situation
that leaves threads spin on an empty queue.

GitOrigin-RevId: d95a2ab5254fb59ddb33c55dcfb89ab395fe6a2d