[tests] generate the runtime module repository in tests which use it (IJPL-201730)
If Bazel compiles the modules, it won't generate the runtime module repository. This will be done in the build scripts, but if tests are started from the IDE this should be done inside tests. So  `generateRuntimeModuleRepositoryForTests` function is introduced which is used to provide the repository for tests which use it.

IJ-MR-174028

GitOrigin-RevId: 717abb92cd7de788b7a5ca5c50defb028a9a088d