[remote dev] pass the path to the IDE home explicitly when the embedded client is started without launcher (IJPL-201730)
If the IDE is started not from a production distribution, but with classes packed to JARs, it won't be able to use the default launcher to run the embedded client. So we need to pass required properties to the Java process explicitly to ensure that PathManager.getHomeDir inside the embedded client works.

IJ-MR-174028

GitOrigin-RevId: 6d7a71d8973d8e653cbe2e1a3896ef036a790890