[build scripts] add a separate flag to remember that output for all modules was generated (IJPL-201730)
It's not correct to set 'useCompiledClassesFromProjectOutput' to 'true' if module outputs aren't unpacked to the classes output directory. This led to a warning about missing production output directory if 'reuseOrCompile' is called again, and will lead to an error about missing whole classed output directory if the runtime module repository isn't generated during compilation. So now a separate 'outputForAllModulesIsAvailable' flag is introduced, and it's used to skip compilation.

IJ-MR-174028

GitOrigin-RevId: 15160e8e3dbdcaed1d0b0581b8cd420798011970