[JEWEL-997] Write Maven artefacts validation script
This adds a new script, validate-maven-artifacts.main.kts, that ensures that the
Jewel artefacts publishing is set up correctly. The script will flag:
 * Differences in JewelMavenArtifacts across branches
 * Artefacts not being published in one or more branches
 * Artefact dependencies being different across branches

Usage: validate-maven-artifacts [<options>] <branches>...

This also fixes the usage help and readme for jps-bootstrap to mention that you
need to pass the project root (the ultimate or community root path) to the script.
There was no mention of it and I was having issues with the script, so after
checking with Nikolay Chashnikov, we agreed that should be updated. I also tried
asking Leonid Shalupov, just in case, but he has not replied to my messages.