[rdct] IJPL-197499: Restore active app logic when requesting focus
In 89424d17824276b2532123cbc63db5ed094f3d3a a project leak was fixed by checking if the active window was disposed before returning it. However, since getActiveWindow was now used in focusProjectWindow too, the app could be considered inactive if the active window was a disposed modal progress indicator, for example. This led to focus sometimes not being requested on Linux with the X Window System. To restore the original logic, I removed the check for a disposed window and checked whether the active project itself is disposed in getActiveProject (an alternative solution mentioned by Sergei Tachenov in the original commit).

GitOrigin-RevId: 3f6ab9c6f887f26e889fe1996b071d26297ce21b