IJPL-205256 Introduce `default.language.level.name` and `default.project.jdk.name` as a fallback for default sdk configuration
ProjectSdk and java LanguageLevel are now moved from project configuration to WSM. This means that the DefaultProject does not contain this configuration anymore. We hope that no one actually need it, because all the wizards should have its own configuration to select the SDK. New registry values are introduced as a backup plan if we overlooked some important use cases.

GitOrigin-RevId: 58ced9c5e25b66e9dc992af84afcbd63ef42baf0