[gradle] IDEA-377941 remove the relation between GradleProjectEntity and GradleBuildEntity
Although `GradleContentRootSyncContributorTest` runs with no issues, there are some tests that are failing because of this relation:
* com.intellij.compose.ide.plugin.resources.ComposeResourcesGotoDeclarationTest
* org.jetbrains.kotlin.idea.k2.codeInsight.gradle.KotlinGradleCompletionTest
* and others, see https://patronus.labs.jb.gg/robot/0770deac-67ea-416e-bd0c-c7f63997fb5a

The possible cause: one of the sync phases runs with no changes. The workspace model tries replacing the existing entities with the same ones. But traverses them in a wrong order, so decides that the parent is not created yet. Tries creating it but fails because this entity was not prepared for creation. This causes an issue like "NoSuchElementException: Key 10 is missing in the map."

This issue was described also in IJPL-204027, and it has a workaround described there. I tried applying it to the relation between GradleBuildEntity and GradleProjectEntity, but it didn't fix the tests. So I just removed this relation because GradleProjectEntity also has buildId - it's possible to find a related build if needed.

I've already seen this error before, I described it in IJPL-193757. That time I also "solved" the problem by avoiding creation the relationship between entities.

Tests:
* GradleContentRootSyncContributor

GitOrigin-RevId: d6f67af87dca28dbba268c974b621bd57778baec