[platform] refactoring: make some members used from tests public (IJPL-205216)
This is needed to avoid IllegalAccessError if tests are loaded by a different classloader and to allow enabling 'Suspicious Package-Private Access' for tests as well.

Visibility for some `protected` members is changed to `public` where it looks reasonable, in other places separate accessor methods for test are added. In some cases the problems are fixed by using getters instead of accessing protected fields directly.

GitOrigin-RevId: 378e5cba73f9b99714e1b7dfd7040ed53fbaad52