[terminal] IJPL-205317 Remove the copy provider from the terminal data context
The copy action is a special action: it works even without an editor
in the data context because BasicUiDataRule looks for an editor
directly in the Swing hierarchy.

But because it's a rule, any component can override the value
of the copy provider, as components have a higher priority
in the data context than rules.

So by setting the copy provider to null we effectively disable
the copy action for good. Our own copy/paste actions will
continue working, as they don't rely on providers.

GitOrigin-RevId: 92b9184822fd357ba4337d0c8ab61bf53fd2b16b